examples: we need kokkos to be installed
It was decided to add Kokkos within a new image (built on top of the
images we alrady have).

The use cases would go in the "examples" directory, and we need to
pass `ReProspect_ENABLE_EXAMPLES` to compile them.

We can compile Kokkos for a single GPU architecture, so we decided to only
compile it for the architectures of our runners, since we'll probably only run tests
that actually need to run the executables.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>